Fixed setting subscriptions in review details section
